oms-monolith/api/graphql/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.215.1
httpx==0.25.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
nats-py==2.6.0
websockets==12.0